requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-login
google-generativeai
python-docx
pypdf
markdown
gunicorn